requirements.txt
python-dotenv==1.1.1
pandas==2.3.3
numpy==2.3.3
scikit-learn==1.7.2
pyarrow==21.0.0
torch==2.8.0
xgboost==3.0.5
matplotlib==3.10.6
fastapi==0.110.0
uvicorn[standard]==0.37.0
gunicorn==23.0.0
pydantic==2.11.10
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
httpx==0.28.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
requests==2.32.5
Flask==3.1.2